Update Oblique Strategies.html
Change the code so if you click anywhere it will randomly generate a oblique strategy.. Also made it so that the sentence is in the center of the screen.. No buttons and no header text. When you open the site it now says "inspire" in the center of the screen. Then after clicking anywhere it'll generate a random strategy. Also no strategy is to be repeated until the user has clicked through all the strategies.